requirements.txt
requests
dnspython
python-whois
coloroma